Standardize provider file layout: move Hue + PlayStation into subfolders (v4.1.32)
LIFX already lived under Devices/LIFX/ with its provider co-located with
the rest of its files. Hue and PlayStation had their providers at the
Devices/ root level instead (Hue: namespace was already .Hue but the
file was in the wrong folder; PlayStation: namespace was the parent
.Devices). Moving both providers into their dedicated subfolders + the
PlayStation namespace correction so all three custom providers follow
the same shape:

  Devices/Hue/HueRGBDeviceProvider.cs           (namespace already .Hue)
  Devices/PlayStation/PlayStationControllerRGBDeviceProvider.cs
                                                (namespace .Devices -> .PlayStation)
  Devices/LIFX/LifxRGBDeviceProvider.cs         (already in place)

Updated three call sites that resolved PlayStationControllerRGBDeviceProvider
through the parent namespace to add a `using ...PlayStation;`:
RGBController, SettingsViewModel, DecoratorHarnessUI.MainViewModel.

Used `git mv` so source history follows the file. No behaviour change;
build clean, 130/130 tests pass.
