+++ Chromatics/Core/RGBController.cs
@@ -2,6 +2,7 @@
 using Chromatics.Extensions.RGB.NET.Devices;
 using Chromatics.Extensions.RGB.NET.Devices.Hue;
 using Chromatics.Extensions.RGB.NET.Devices.LIFX;
+using Chromatics.Extensions.RGB.NET.Devices.PlayStation;
 using Chromatics.Helpers;
 using Chromatics.Layers;
 using Chromatics.Models;
