+++ Chromatics/Extensions/RGB.NET/Devices/Hue/HueRGBDeviceProvider.cs
